--- src/classes/table/ColumnType.ts
@@ -1,11 +1,0 @@
-export enum ColumnType {
-  UNDEFINED = 0,
-  CITY,
-  BUILDING,
-  ROOM,
-  NAME,
-  STATUS,
-  LOCATION,
-  DATE,
-  PERSON
-}
