--- src/classes/table/TableCell.ts
@@ -1,7 +1,0 @@
-import  { ColumnType } from "@/classes/table/ColumnType";
-
-export interface TableCell {
-  id: string,
-  displayName: string,
-  type: ColumnType
-}
