+++ src/components/locationRegistration/AddRoom.vue
@@ -53,9 +53,9 @@ import InputField from "@/components/standardUi/InputField.vue";
 import SmallBtnFinish from "@/components/standardUi/SmallBtnFinish.vue";
 import RoomRequest from "@/classes/requests/RoomRequest";
 import CBSearchSuggestion from "@/components/standardUi/CBSearchSuggestions.vue";
-import { roomService } from "@/services/locatieService/roomservice";
-import Building from "@/classes/Building";
-import { buildingService } from "@/services/locatieService/buildingservice";
+import { roomService } from "@/location/roomservice";
+import Building from "@/location/Building";
+import { buildingService } from "@/location/buildingservice";
 import LinkOrStayModal from "@/components/standardUi/LinkOrStayModal.vue";
 import { getCurrentInstance } from "@vue/runtime-core";
 import SelectOption from "@/classes/helpers/SelectOption";
