+++ src/location/components/LocationProxy.vue
@@ -6,22 +6,22 @@
 </template>
 
 <script lang="ts">
-    import { Vue, Options } from 'vue-class-component'
+    import { Vue } from 'vue-class-component'
     import { Prop } from 'vue-property-decorator'
-    import { personeelService } from '../../services/personeelService/personeelService';
+    import { roomService } from '@/location/roomservice'
 
-    export default class ReceiverProxy extends Vue {
+    export default class LocationProxy extends Vue {
         @Prop() id!: string;
         private isLoaded: boolean = false;
         private text: string = "Empty";
 
         async mounted() {
-            personeelService.get(this.id)
+            roomService.getById(this.id)
                 .then((res) => {
                     this.text = res.name;
                     this.isLoaded = true;
                 })
-                .catch ((err) =>
+                .catch((err) =>
                     console.error(err)
                 );
         }
@@ -30,8 +30,13 @@
 
 <style scoped lang="scss">
     @keyframes fade-anim {
-        from{background-color: lightgray;}
-        to{background-color: darkgray;}
+        from {
+            background-color: lightgray;
+        }
+
+        to {
+            background-color: darkgray;
+        }
     }
 
     .fade {
