+++ src/package/Package.ts
@@ -1,7 +1,5 @@
 import ISerializable from "@/classes/helpers/ISerializable"
-import Ticket from "@/classes/Ticket";
-import { Person } from "@/classes/Person";
-import { personeelService } from "@/services/personeelService/personeelService"
+import Ticket from "@/package/Ticket";
 
 export class Package {
     id: string;
@@ -37,18 +35,6 @@ export class Package {
         );
     }
 
-    //deserialize(data: Object): Package {
-    //    this.id = data["id"];
-    //    this.receiverId = data["receiverId"];
-    //    this.collectionPointId = data["collectionPointId"];
-    //    this.sender = data["sender"];
-    //    this.status = data["status"];
-    //    this.routeFinished = data["routeFinished"];
-    //    this.tickets = data["tickets"];
-    //    this.name = data["name"];
-    //    return this;
-    //}
-
     getFirstTicket(): Ticket | null {
         const lastIndex = this.tickets.length - 1;
         if (lastIndex < 0) return null
@@ -62,10 +48,5 @@ export class Package {
 
     getStatus(): string {
         return this.routeFinished ? "Afgeleverd" : "In behandeling";
-    }
-
-    async getReceiver() {
-        if (this.receiverId)
-        return await personeelService.get(this.receiverId);
     }
 }
